prefer linking matching parent instead of all children
when using automated linking based on the link map configuration, the
current behavior seem to prefer linking each of the children instead of
the parent when both match the same pattern, i.e. instead of multiple
links we should only link the parent (or perhaps the first match).

this should reduce the number of links created, especially if the
parent contain multiple children.

closes #7