suppress warning for passing registry password in cli
the `docker login` command is complaining that we pass the password via
a cli argument instead of using the `--password-stdin`.

to suppress this warning we write the passwords to separate files and
pipe it to the command using stdin when we login.

the security issues should be the same when we run it in the pipeline,
as long as we ensure that we remove the files in `after_script`.